refactor(tests): extract `FeedLoader` test helpers into shared protocol extension to remove duplication
